+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+## [v2.0.1](https://github.com/PolymerLabs/dom5/tree/v2.0.1) (2016-11-01)
+- TypeScript type definition fixes
+
 ## [v2.0.0](https://github.com/PolymerLabs/dom5/tree/v2.0.0) (2016-09-23)
 - Upgraded to `parse5@^v2.2.1`.
 - (*breaking*) Because parse5 2.x correctly handles `<template>` elements,